fix: 소설 평가 타입 변경

+++ app/src/main/java/com/teamwss/websoso/data/model/MyActivitiesEntity.kt
@@ -19,7 +19,7 @@ data class MyActivitiesEntity(
         val novelId: Int?,
         val title: String?,
         val novelRatingCount: Int?,
-        val novelRating: Double?,
+        val novelRating: Float?,
         val relevantCategories: List<String>,
     )
 }
